% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{ComponentPlot}
\alias{ComponentPlot}
\title{ComponentPlot}
\usage{
ComponentPlot(
  data,
  identity = "orig.ident",
  component = "DKCC",
  feature = "MALAT1",
  show.unassigned = TRUE,
  show.pct = FALSE,
  show.gene.exp = FALSE,
  do.label = T
)
}
\arguments{
\item{data}{seurat object}

\item{identity}{column name in metadata for sample identity}

\item{component}{column name in metadata for annotation/classification identity}

\item{feature}{gene for expression visualisation, only shown if show.gene.exp is TRUE}

\item{show.unassigned}{include unassigned cells in visualisation, TRUE by default}

\item{show.pct}{change plot to show percentage of cells instead of total cell number}

\item{show.gene.exp}{colour by expression of "feature" in each annotation category}

\item{do.label}{add label to plot, TRUE by default}
}
\value{
ggplot object
}
\description{
ComponentPlot
}
\examples{
ComponentPlot(organoid, show.pct = TRUE)
}
