% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevKidCC_Models.R
\docType{data}
\name{DevKidCC_Models}
\alias{DevKidCC_Models}
\alias{model1.all}
\alias{model2.nephron}
\alias{model2.stroma}
\alias{model2.urep}
\alias{model3.dn}
\alias{model3.pn}
\alias{model3.rc}
\alias{reference}
\title{scPred Models}
\format{
An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{scPred} of length 1.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 843696 rows and 11 columns.
}
\usage{
model1.all

model2.nephron

model2.stroma

model2.urep

model3.dn

model3.pn

model3.rc

reference
}
\description{
These are the models trained by scPred on the reference data
}
\keyword{datasets}
