% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKCC.R
\name{DKCC}
\alias{DKCC}
\title{DKCC}
\usage{
DKCC(seurat, threshold = 0.7, max.iter = 1)
}
\arguments{
\item{seurat}{seurat object}

\item{threshold}{minimum value for an identity to be assigned within the model call, default is 0.7}

\item{max.iter}{Can ask scPred to run this number of integrations, set to 0 be default}
}
\value{
seurat object with additional metadata columns
}
\description{
DKCC
}
\examples{
organoid <- DKCC(organoid)
}
