% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_summary.R
\name{GeneSummary}
\alias{GeneSummary}
\title{GeneSummary}
\usage{
GeneSummary(
  data,
  identity = "orig.ident",
  split.by = "DKCC",
  features = rownames(data),
  cells = colnames(data),
  do.norm = FALSE,
  group.by = NULL
)
}
\arguments{
\item{data}{seurat object}

\item{identity}{column name in metadata describing required sample identity}

\item{split.by}{column name in metadat describing required annotation/classification identity}

\item{features}{features to use, defaults to all}

\item{cells}{cells to use, defaults to all}

\item{do.norm}{option to run normalisation on data if not already done, set TRUE if required}

\item{group.by}{secondary grouping identity, default is NULL}
}
\description{
GeneSummary
}
\examples{
output <- GeneSummary(organoid, features = c("GAPDH", "SIX2", "JAG1", "EPCAM", "NPHS1"))
}
