% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{GetSampleIDs}
\alias{GetSampleIDs}
\title{GetSampleIDs}
\usage{
GetSampleIDs(
  sample = NULL,
  reference = NULL,
  type = NULL,
  base_protocol = NULL,
  modification = NULL,
  age = NULL,
  union = FALSE
)
}
\arguments{
\item{sample}{Not required}

\item{reference}{Original publication reference eg. Howden_2019}

\item{type}{can select either "Organoid" or "iUB" if required}

\item{base_protocol}{Select base protocol/s of interest, including "Takasato", "Morizane", "Freedman", "Kumar", "Low", "Howden-Wilson", "Mae"}

\item{modification}{TRUE/FALSE for if that base protocol was modified in some way}

\item{age}{age range of organoids, string of integers eg. 1:18}

\item{union}{if multiple fields are entered, use union=TRUE to select all samples or union=FALSE to select only the intersect i.e. samples selected from all fields}
}
\value{
string of sample IDs that can be input into DotPlotCompare.
}
\description{
GetSampleIDs
}
