% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{IdentMeans}
\alias{IdentMeans}
\title{IdentMeans}
\usage{
IdentMeans(
  data,
  group = "protocol",
  identity = "orig.ident",
  component = "DKCC"
)
}
\arguments{
\item{data}{seurat object with DevKidCC annotation}

\item{group}{identity to generate mean scores from - x axis}

\item{identity}{individual samples}

\item{component}{class of annotation of interest}
}
\value{

}
\description{
IdentMeans
}
