% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organoid.R
\docType{data}
\name{organoid}
\alias{organoid}
\title{Curated Seurat object for Howden et al. 2018 which with DevKidCC classification}
\format{
Seurat object contains data from two time points. Only top 2000 variable genes included in sample dataset.
\describe{
\item{age}{age of the organoids}
\item{Identity}{Identity of original annotation}
...
}
}
\source{
\url{https://doi.org/10.15252/embr.201847483}
}
\usage{
organoid
}
\description{
Curated Seurat object for Howden et al. 2018 which with DevKidCC classification
}
\keyword{datasets}
