% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{SankeyPlot}
\alias{SankeyPlot}
\title{SankeyPlot}
\usage{
SankeyPlot(
  data,
  simple = T,
  origin = "sample.id",
  end = "DKCC",
  height = 900,
  width = 1200,
  fontsize = 12,
  sinksRight = TRUE,
  LinkGroup = "Var1",
  remove.unassigned = F,
  nephron.only = F
)
}
\arguments{
\item{data}{seurat object fot the data to be pulled from}

\item{simple}{defualt is FALSE; TRUE changes to a direct plot going from the origin to final classification without intermediate steps shown}

\item{origin}{the meta.data column describing the origin identity}

\item{height}{the height of the plot}

\item{width}{the width of the plot}

\item{fontsize}{the fontsize of the plot}

\item{sinksRight}{default FALSE; graphing parameter from base plotting function that forces final identity to right side of graph}

\item{LinkGroup}{Changes the way in which linking colours are plotted. Default is "Var3" representing the origin population}

\item{remove.unassigned}{default FALSE; determines if unassigned values are removed from the plot}

\item{nephron.only}{default FALSE; TRUE will plot only cell types that are classified into a nephron lineage}

\item{forced}{default FALSE IF TRUE changes the outcome to show all cells force called into a cell type instead of being unassigned}
}
\value{
the plot
}
\description{
Plots a Sankey diagram shown how the cells are from a category into the final classification
}
