% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{ComponentPlot}
\alias{ComponentPlot}
\title{ComponentPlot}
\usage{
ComponentPlot(
  data,
  identity = "orig.ident",
  component = "DKCC",
  feature = "MALAT1",
  show.unassigned = TRUE,
  show.pct = FALSE,
  show.gene.exp = FALSE,
  do.label = T
)
}
\arguments{
\item{do.label}{}
}
\value{

}
\description{
ComponentPlot
}
