% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevKidCC-package.R
\docType{package}
\name{DevKidCC-package}
\alias{DevKidCC}
\alias{DevKidCC-package}
\title{DevKidCC: Kidney Cell Classifier}
\description{
Thsi package takes a Seurat based single cell dataset and classifies the cells into their most likely kidney cell type.
}
\author{
\strong{Maintainer}: Sean Wilson \email{sean.wilson@mcri.edu.au} (\href{https://orcid.org/0000-0002-8994-0781}{ORCID})

}
\keyword{internal}
