% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{DotPlotCompare}
\alias{DotPlotCompare}
\title{DotPlotCompare}
\usage{
DotPlotCompare(
  new.object = NULL,
  show = "gene",
  split.by = "orig.ident",
  scaling = "log",
  size = "pct.exp",
  idents = "all",
  classification = "DKCC",
  dot.min = 0.1,
  dot.scale = 6,
  col.min = NA,
  col.max = NA,
  scale.min = NA,
  scale.max = NA,
  scale.by = "radius",
  features,
  compare.to.organoids = FALSE,
  filter.samples = NULL,
  columns = NULL
)
}
\arguments{
\item{new.object}{seurat object to compare to reference data}

\item{show}{this defines the focus of the output; "gene" partitions by gene, anything else partitions by segment}

\item{split.by}{meta.data column to split the sample by, usually the sample column}

\item{scaling}{choose "log", "scale" or "raw" to change how the expression values are displayed}

\item{size}{how to scale the size of the dot, "pct.exp" for the percent of cells expressing the gene, "Pct" is percent of cells in identity}

\item{idents}{Identities to display. Use "all" to show all segments, "nephron" to show all nephron segments, "others" to show all non-nephron segments, or a string of identities to select specific Identities}

\item{classification}{choose "DKCC" or "LineageID"}

\item{dot.min}{dot size of minimum cell percentage}

\item{dot.scale}{dot size of maximum cell percentage}

\item{col.min}{minimum expression value to plot}

\item{col.max}{maximum expression value to plot}

\item{scale.min}{minimum scale}

\item{scale.max}{maximum scale}

\item{scale.by}{"radius" or "size"}

\item{features}{features to plot}

\item{compare.to.organoids}{if function should plot organoid samples from the database}

\item{filter.samples}{if compare.to.organoids=T, can provide desired sample names here using the output from "GetSampleIDs".}

\item{columns}{how many columns to plot. Default is to plot as square as possible.}
}
\value{
ggplot object
}
\description{
DotPlotCompare
}
