% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_summary.R
\name{GeneSummary}
\alias{GeneSummary}
\title{GeneSummary}
\usage{
GeneSummary(
  data,
  identity = "orig.ident",
  split.by = "DKCC",
  features = rownames(data),
  cells = colnames(data),
  group.by = NULL
)
}
\arguments{
\item{group.by}{}
}
\value{

}
\description{
GeneSummary
}
