% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{IdentBoxPlot}
\alias{IdentBoxPlot}
\title{IdentBoxPlot}
\usage{
IdentBoxPlot(
  data,
  group,
  identity = "orig.ident",
  component = "DKCC",
  feature = "MALAT1",
  column = T,
  show.unassigned = TRUE,
  show.pct = FALSE,
  show.gene.exp = FALSE,
  do.label = T
)
}
\arguments{
\item{do.label}{}
}
\value{

}
\description{
IdentBoxPlot
}
